+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 0.3.17 [unreleased]
+- feat: Implement Keystore [PR 74]
 - feat: Added Ipfs::remove_peer and Ipfs::remove_peer_address [PR 73]
 - feat: Implements MultiaddrExt and remove wrapper [PR 72]
 - feat: Basic AddressBook implementation [PR 71]
@@ -12,6 +13,7 @@
 [PR 71]: https://github.com/dariusc93/rust-ipfs/pull/71
 [PR 72]: https://github.com/dariusc93/rust-ipfs/pull/72
 [PR 73]: https://github.com/dariusc93/rust-ipfs/pull/73
+[PR 74]: https://github.com/dariusc93/rust-ipfs/pull/74
 
 # 0.3.16
 - fix: Return events from gossipsub stream [PR 68]